Added Patient and Doctor classes inheritting from User superclass

+++ Snapshot.cpp
@@ -2,7 +2,7 @@
 #include "class.hpp"
 
 // --- Set functions --- //
-void Snapshot::set_id(int id) {
+void Snapshot::set_id(size_t id) {
 	this->id = id ;
 }
 void Snapshot::set_path(std::string path) {
@@ -10,7 +10,7 @@ void Snapshot::set_path(std::string path) {
 }
 
 // --- Get functions --- //
-int Snapshot::get_id() {
+size_t Snapshot::get_id() {
 	return this->id ;
 }
 std::string Snapshot::get_path() {
